Start scaffolding a buffering connection
Based on Akka sample, per comment. Calls the buffer method to build up
the storage, and the acknowledge method tears down the storage. This
doesn't really work and needs a lot of work, since it was based on an
echo server (and stupidly a lot of the buffering was built with this in
mind, in the example).

Honestly, who couples buffering so closely with echoing when writing an
example of buffering?